http: simplify client builder
Simplify the client builder. The client builder originally involved 4
structs: the client, client builder, config, and config builder. The
client builder was a wrapper on top of the config builder, which would
build a config that the client builder would use to configure a client
upon building.

Separately using the config builder to create a config would have no
purpose, since there was no implementation to create a client from a
config; you'd always have to go through the client builder to make use
of the config.

This is complicated, so just simplify the whole thing down to the client
and the client builder.

Closes #215.
